+++ src/TodoForm.jsx
@@ -1,0 +1,12 @@
+/*TodoFrom component that returns a form*/
+function TodoForm() {
+  return (
+    <form>
+      <label htmlFor="todoTitle">Todo:</label>
+      <input id="todoTitle"></input>
+      <button>Add Todo</button>
+    </form>
+  );
+}
+
+export default TodoForm;
